+++ clients/admin-ui/src/features/system/SystemFormTabs.tsx
@@ -8,8 +8,9 @@ import { useAppDispatch, useAppSelector } from "~/app/hooks";
 import DataTabs, { type TabData } from "~/features/common/DataTabs";
 import { useSystemOrDatamapRoute } from "~/features/common/hooks/useSystemOrDatamapRoute";
 import { DEFAULT_TOAST_PARAMS } from "~/features/common/toast";
+import ConnectionForm from "~/features/datastore-connections/system_portal_config/ConnectionForm";
 import PrivacyDeclarationStep from "~/features/system/privacy-declarations/PrivacyDeclarationStep";
-import { System } from "~/types/api";
+import { System, SystemResponse } from "~/types/api";
 
 import { selectActiveSystem, setActiveSystem } from "./system.slice";
 import SystemInformationForm from "./SystemInformationForm";
@@ -76,7 +77,7 @@ const SystemFormTabs = ({
   const router = useRouter();
   const toast = useToast();
   const dispatch = useAppDispatch();
-  const activeSystem = useAppSelector(selectActiveSystem);
+  const activeSystem = useAppSelector(selectActiveSystem) as SystemResponse;
 
   const handleSuccess = (system: System) => {
     // show a save message if this is the first time the system was saved
@@ -124,7 +125,9 @@ const SystemFormTabs = ({
     if (
       index === 0 ||
       (index === 1 && tabIndex === 2) ||
-      (index === 2 && tabIndex === 1)
+      (index === 2 && tabIndex === 1) ||
+      index === 3 ||
+      tabIndex === 3
     ) {
       setTabIndex(index);
     } else {
@@ -209,6 +212,28 @@ const SystemFormTabs = ({
             </Text>
           </Box>
           <DataFlowAccordion system={activeSystem} isSystemTab />
+        </Box>
+      ) : null,
+      isDisabled: !activeSystem,
+    },
+    {
+      label: "Integrations",
+      content: activeSystem ? (
+        <Box width={{ base: "100%", lg: "70%" }}>
+          <Box px={6} paddingBottom={2}>
+            <Text fontSize="sm" lineHeight={5} fontWeight="medium">
+              Integrations are used to process privacy requests like access,
+              erasure, portability, rectification, and consent. Many common
+              systems have API integrations that we can use to automatically
+              process privacy requests. However, email and manual integrations
+              are also available for the systems & functions where API
+              integrations are not available or the preference.
+            </Text>
+          </Box>
+          <ConnectionForm
+            connectionConfig={activeSystem.connection_configs}
+            systemFidesKey={activeSystem.fides_key}
+          />
         </Box>
       ) : null,
       isDisabled: !activeSystem,