+++ src/fides/api/ctl/routes/system.py
@@ -42,7 +42,6 @@ from fides.api.ctl.database.system import (
 )
 from fides.api.ctl.schemas.system import SystemResponse
 from fides.api.ctl.sql_models import System  # type: ignore[attr-defined]
-from fides.api.ctl.utils.api_router import APIRouter
 from fides.api.models.connectionconfig import ConnectionConfig
 from fides.api.oauth.utils import verify_oauth_client, verify_oauth_client_prod
 from fides.api.schemas.connection_configuration.connection_config import (
@@ -52,6 +51,7 @@ from fides.api.schemas.connection_configuration.connection_config import (
     SaasConnectionTemplateResponse,
     SaasConnectionTemplateValues,
 )
+from fides.api.util.api_router import APIRouter
 from fides.api.util.connection_util import patch_connection_configs
 from fides.api.util.system_manager_oauth_util import (
     verify_oauth_client_for_system_from_fides_key_cli,